AudioContext based on input sample rate
Create our AudioContext based on the input sample rate, and re-create it if
that sample rate changes. This will force us to use libav's resampler, instead
of the operating system's resampler, which I'm hoping is the problem causing
clicks and pops on Apple devices. However, this doesn't work on Safari.